Extract base class and format
