+++ tests/PySys/apt-plugin/apt-install-dense/run.py
@@ -1,38 +1,18 @@
-import pysys
-from pysys.basetest import BaseTest
+import sys
+
+sys.path.append("apt-plugin")
+from environment_apt_plugin import AptPlugin
 
 """
 Validate apt plugin install
 
 Using `rolldice` as a guinea pig: [small and without impacts](https://askubuntu.com/questions/422362/very-small-package-for-apt-get-experimentation)
+
+When we list all packages
+When we install a package
+Then we find the package in the list of installed packages
 """
 
-class AptPlugin(BaseTest):
-
-    def setup(self):
-        self.apt_plugin = "/etc/tedge/sm-plugins/apt"
-        self.apt_get = "/usr/bin/apt-get"
-        self.sudo = "/usr/bin/sudo"
-
-
-    def plugin_cmd( self, command, outputfile, exit_code, argument=None):
-        args=[self.apt_plugin, command]
-        if argument:
-            args.append(argument)
-
-        process = self.startProcess(
-            command=self.sudo,
-            arguments=args,
-            stdouterr=outputfile,
-            expectedExitStatus=f"=={exit_code}",
-        )
-
-    def apt_remove(self, package):
-            self.startProcess(
-            command=self.sudo,
-            arguments=[self.apt_get, 'remove', '-y', package],
-            abortOnError=False,
-        )
 
 class AptPluginInstallTest(AptPlugin):
     def setup(self):
@@ -41,13 +21,13 @@ class AptPluginInstallTest(AptPlugin):
         self.addCleanupFunction(self.remove_rolldice_module)
 
     def execute(self):
-        self.plugin_cmd('list', 'outp_before', 0)
-        self.plugin_cmd('install', 'outp_install', 0, "rolldice")
-        self.plugin_cmd('list', 'outp_after', 0)
+        self.plugin_cmd("list", "outp_before", 0)
+        self.plugin_cmd("install", "outp_install", 0, "rolldice")
+        self.plugin_cmd("list", "outp_after", 0)
 
     def validate(self):
-        self.assertGrep ("outp_before.out", 'rolldice', contains=False)
-        self.assertGrep ("outp_after.out", 'rolldice', contains=True)
+        self.assertGrep("outp_before.out", "rolldice", contains=False)
+        self.assertGrep("outp_after.out", "rolldice", contains=True)
 
     def remove_rolldice_module(self):
-        self.apt_remove('rolldice')
+        self.apt_remove("rolldice")
